backend/requirements.txt
fastapi
uvicorn
mariadb
python-dotenv
pydantic-settings 
pydantic[email]
bcrypt
pyjwt
passlib
